Relocate ConsoleLogger to plugins
