--- src/common/logger/index.ts
@@ -1,2 +1,0 @@
-export * from './logger';
-export * from './consoleLogger';
