+++ src/plugins/loggers/index.ts
@@ -1,0 +1,2 @@
+export * from './logger';
+export * from './ConsoleLogger';
